backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.4
snowflake-connector-python==3.10.0
pydantic==2.8.2
python-dotenv==1.0.1


